+++ projects/wind-daq/apps/desktop-wails/frontend/bindings/wind-daq/services/api-go/internal/core/device/index.js
@@ -6,6 +6,7 @@ export {
     ChannelConfig,
     ChannelSensorType,
     Connection,
+    DaqT1602HardwareConfig,
     DaqT1603HardwareConfig,
     DataPayload,
     Profile,